Fix undefined behaviour in test

+++ tests/test-output-flex-way-change.cpp
@@ -44,8 +44,8 @@ TEST_CASE("change way from t1")
 
     options.append = true;
 
-    unsigned long num_t1 = 0;
-    unsigned long num_t2 = 0;
+    int num_t1 = 0;
+    int num_t2 = 0;
     std::string update;
     SECTION("to t2")
     {
@@ -87,8 +87,8 @@ TEST_CASE("change way from t2")
 
     options.append = true;
 
-    unsigned long num_t1 = 0;
-    unsigned long num_t2 = 0;
+    int num_t1 = 0;
+    int num_t2 = 0;
     std::string update;
     SECTION("to t1")
     {
@@ -131,8 +131,8 @@ TEST_CASE("change way from t1 and t2")
 
     options.append = true;
 
-    unsigned long num_t1 = 0;
-    unsigned long num_t2 = 0;
+    int num_t1 = 0;
+    int num_t2 = 0;
     std::string update;
     SECTION("to t1")
     {
